feat: integration AI plugin by NX generator
