+++ packages/plugins/integration-ai/.gitignore
@@ -1,0 +1,6 @@
+# dependencies
+node_modules/
+
+# misc
+npm-debug.log
+dist
