+++ packages/plugins/integration-ai/.npmignore
@@ -1,0 +1,4 @@
+# .npmignore
+
+src/
+node_modules/
