+++ packages/plugins/integration-ai/README.md
@@ -1,6 +1,14 @@
 # Integration Gauzy AI
 
+This library was generated with [Nx](https://nx.dev).
+
+# Description
+
 This library provides integration with Gauzy AI project.
+
+## Building
+
+Run `yarn run build` to build the library.
 
 ## Running unit tests
