--- packages/plugins/integration-ai/jest.config.js
@@ -1,9 +1,0 @@
-module.exports = {
-	name: 'integration-ai',
-	preset: '../../../jest.config.js',
-	transform: {
-		'^.+\\.[tj]sx?$': 'ts-jest'
-	},
-	moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx', 'html'],
-	coverageDirectory: '../../coverage/plugins/integration-ai'
-};
