+++ packages/plugins/integration-ai/jest.config.ts
@@ -1,0 +1,11 @@
+/* eslint-disable */
+export default {
+	displayName: 'integration-ai',
+	preset: '../../../jest.preset.js',
+	testEnvironment: 'node',
+	transform: {
+		'^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }]
+	},
+	moduleFileExtensions: ['ts', 'js', 'html'],
+	coverageDirectory: '../../../coverage/packages/plugins/integration-ai'
+};
