+++ packages/plugins/integration-ai/src/index.ts
@@ -1,4 +1,4 @@
-export * from './sdk/gauzy-ai-sdk';
-export * from './gauzy-ai.service';
-export * from './gauzy-ai.module';
-export * from './request-config.provider';
+export * from './lib/sdk/gauzy-ai-sdk';
+export * from './lib/gauzy-ai.service';
+export * from './lib/gauzy-ai.module';
+export * from './lib/request-config.provider';
