+++ packages/plugins/integration-ai/src/lib/config/gauzy-ai.ts
