+++ packages/plugins/integration-ai/src/lib/configuration.interface.ts
@@ -1,15 +1,14 @@
-
 // IApiKeyMethod represents an interface for ApiKey authentication method, with optional ApiKey and ApiSecret properties.
 export interface IApiKeyMethod {
-    apiKey?: string;
-    apiSecret?: string;
-    openAiSecretKey?: string;
-    openAiOrganizationId?: string;
+	apiKey?: string;
+	apiSecret?: string;
+	openAiSecretKey?: string;
+	openAiOrganizationId?: string;
 }
 
 export interface IBearerTokenMethod {
-    tenantIdApi?: string;
-    bearerTokenApi?: string;
+	tenantIdApi?: string;
+	bearerTokenApi?: string;
 }
 
 export type IConfigurationOptions = IApiKeyMethod & IBearerTokenMethod;
@@ -18,18 +17,18 @@ export type IConfigurationOptions = IApiKeyMethod & IBearerTokenMethod;
  * Represents common HTTP methods as string values.
  */
 export enum HttpMethodEnum {
-    GET = 'GET',
-    POST = 'POST',
-    PUT = 'PUT',
-    DELETE = 'DELETE',
-    PATCH = 'PATCH',
-    HEAD = 'HEAD',
-    OPTIONS = 'OPTIONS',
+	GET = 'GET',
+	POST = 'POST',
+	PUT = 'PUT',
+	DELETE = 'DELETE',
+	PATCH = 'PATCH',
+	HEAD = 'HEAD',
+	OPTIONS = 'OPTIONS'
 }
 
 /**
  * Represents HTTP request headers as an object where keys are header names and values are header values.
  */
 export interface AxiosRequestHeaders {
-    [key: string]: string;
+	[key: string]: string;
 }
