+++ packages/plugins/integration-ai/src/lib/constants.ts
