+++ packages/plugins/integration-ai/src/lib/sdk/queries/query.graphql
