+++ packages/plugins/integration-ai/src/lib/sdk/schema.graphql
