+++ packages/plugins/integration-ai/tsconfig.lib.json
@@ -1,3 +1,14 @@
 {
-	"extends": "./tsconfig.json"
+	"extends": "./tsconfig.json",
+	"compilerOptions": {
+		"declaration": true,
+		"sourceMap": true,
+		"baseUrl": "./src",
+		"rootDir": "./src",
+		"outDir": "./dist",
+		"types": ["node", "jest"],
+		"target": "es6"
+	},
+	"include": ["src/**/*.ts"],
+	"exclude": ["jest.config.ts", "src/**/*.spec.ts", "src/**/*.test.ts"]
 }
