+++ packages/plugins/sentry-tracing/package.json
@@ -38,6 +38,7 @@
 		"rxjs": "^7.4.0"
 	},
 	"devDependencies": {
+		"@types/jest": "^29.4.4",
 		"@types/node": "^20.14.9",
 		"rimraf": "^3.0.2",
 		"typescript": "5.1.6"
